pyproject.toml
[project]
name = "bigbrotheranalytics"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.11.0",
    "alpha-vantage>=3.0.0",
    "beautifulsoup4>=4.14.2",
    "captum>=0.7.0",
    "dash>=3.2.0",
    "duckdb>=1.4.1",
    "fastapi>=0.121.0",
    "fredapi>=0.5.2",
    "gymnasium>=1.2.2",
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.10",
    "lightgbm>=4.6.0",
    "lime>=0.2.0.1",
    "matplotlib>=3.10.7",
    "nltk>=3.9.2",
    "numpy>=2.3.4",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.7",
    "playwright>=1.55.0",
    "plotly>=6.4.0",
    "polars>=1.35.1",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.4",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "scrapy>=2.13.3",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
    "shap>=0.49.1",
    "spacy>=3.8.8",
    "sqlalchemy>=2.0.44",
    "stable-baselines3>=2.7.0",
    "streamlit>=1.50.0",
    "tabula-py>=2.10.0",
    "tika>=3.1.0",
    "torch",
    "torch-geometric>=2.7.0",
    "torchaudio",
    "torchvision",
    "transformers>=4.57.1",
    "uvicorn>=0.38.0",
    "xgboost>=3.1.1",
    "yfinance>=0.2.66",
]
